+++ src/app/Playground/App.xaml.cs
@@ -24,8 +24,6 @@ public partial class App : Application
             config.AddNavigationButton(string.Empty, "Another button", IconName.bell_line, new Command(() => { }));
             config.AddNavigationButton(string.Empty, "This button?",  IconName.alert_fill, new Command(() => { }));
             config.AddNavigationButton(string.Empty, "No button", IconName.close_line, new Command(() => { }));
-            
-            config.AddPageThatHidesButton(typeof(VetleTestPage1));
         });
     }
 }